requirements.txt
streamlit
langgraph
langchain
requests
matplotlib
pandas
python-dotenv